Feature/deno 1.5 (#517)
* support Deno v1.5.0

* update CHANGELOG

* avoid type-only re-exporting for Deno 1.5

* remove re-exporting NullableOptions

* avoid re-exporting ErrorHandler
