+++ .github/workflows/verify-examples.yml
@@ -191,6 +191,7 @@ jobs:
         - "1.2.0"
         - "1.3.0"
         - "1.4.0"
+        - "1.5.0"
       fail-fast: false
     steps:
     - name: Checkout source codes
