+++ dist-deno/exporter.ts
@@ -1,4 +1,5 @@
 export * from "./libs/applySchemaObject.ts";
+export * from "./libs/publicTypes.ts";
 export * from "./libs/ValueSchemaError.ts";
 export * from "./schemas/array.ts";
 export * from "./schemas/boolean.ts";
