+++ dist-deno/schemaClasses/ArraySchema.ts
@@ -7,7 +7,6 @@ import * as maxLength from "../appliers/array/maxLength.ts";
 import * as minLength from "../appliers/array/minLength.ts";
 import * as type from "../appliers/array/type.ts";
 import { BaseSchema } from "./BaseSchema.ts";
-export { NullableOptions } from "./BaseSchema.ts";
 export type OptionsForArray<T> = converter.Options<T[]> | ifUndefined.Options<T[]> | ifEmptyString.Options<T[]> | ifNull.Options<T[]> | each.Options<T> | minLength.Options | maxLength.Options | type.Options;
 export class ArraySchema<T, Tx = never> extends BaseSchema<T[] | Tx> {
     constructor(options: OptionsForArray<T>) {
