+++ dist-deno/schemaClasses/BooleanSchema.ts
@@ -3,7 +3,6 @@ import * as ifNull from "../appliers/ifNull.ts";
 import * as ifUndefined from "../appliers/ifUndefined.ts";
 import * as type from "../appliers/boolean/type.ts";
 import { BaseSchema } from "./BaseSchema.ts";
-export { NullableOptions } from "./BaseSchema.ts";
 export type OptionsForBoolean = ifUndefined.Options<boolean> | ifEmptyString.Options<boolean> | ifNull.Options<boolean> | type.Options;
 export class BooleanSchema<Tx = never> extends BaseSchema<boolean | Tx> {
     constructor(options: OptionsForBoolean) {
