+++ dist-deno/schemaClasses/EmailSchema.ts
@@ -6,7 +6,6 @@ import * as pattern from "../appliers/email/pattern.ts";
 import * as trims from "../appliers/string/trims.ts";
 import * as type from "../appliers/string/type.ts";
 import { BaseSchema } from "./BaseSchema.ts";
-export { NullableOptions } from "./BaseSchema.ts";
 export type OptionsForEmail = ifEmptyString.Options<string> | ifNull.Options<string> | ifUndefined.Options<string> | pattern.Options | maxLength.Options | trims.Options | type.Options;
 export class EmailSchema<Tx = never> extends BaseSchema<string | Tx> {
     constructor(options: OptionsForEmail) {
