+++ dist-deno/schemaClasses/ObjectSchema.ts
@@ -6,7 +6,6 @@ import * as schema from "../appliers/object/schema.ts";
 import * as type from "../appliers/object/type.ts";
 import { AnyObject, ObjectTypeOf, SchemaObject } from "../libs/types.ts";
 import { BaseSchema } from "./BaseSchema.ts";
-export { NullableOptions } from "./BaseSchema.ts";
 export type OptionsForObject<S> = converter.Options<AnyObject> | ifUndefined.Options<AnyObject> | ifEmptyString.Options<AnyObject> | ifNull.Options<AnyObject> | schema.Options<S> | type.Options;
 export class ObjectSchema<S extends SchemaObject, Tx = never> extends BaseSchema<ObjectTypeOf<S> | Tx> {
     constructor(options: OptionsForObject<S>) {
