+++ dist-deno/schemas/array.ts
@@ -1,4 +1,5 @@
-import { ArraySchema, NullableOptions, OptionsForArray } from "../schemaClasses/ArraySchema.ts";
+import { NullableOptions } from "../libs/publicTypes.ts";
+import { ArraySchema, OptionsForArray } from "../schemaClasses/ArraySchema.ts";
 export function array<T>(options: OptionsForArray<T> & NullableOptions): ArraySchema<T, null>;
 export function array<T>(options: OptionsForArray<T>): ArraySchema<T>;
 export function array<T>(): ArraySchema<T>;
