+++ dist-deno/schemas/number.ts
@@ -1,4 +1,5 @@
-import { NullableOptions, NumberSchema, OptionsForNumber } from "../schemaClasses/NumberSchema.ts";
+import { NullableOptions } from "../libs/publicTypes.ts";
+import { NumberSchema, OptionsForNumber } from "../schemaClasses/NumberSchema.ts";
 export { NUMBER } from "../schemaClasses/NumberSchema.ts";
 export function number(options: OptionsForNumber & NullableOptions): NumberSchema<null>;
 export function number(options: OptionsForNumber): NumberSchema;
