+++ dist-deno/schemas/numericString.ts
@@ -1,4 +1,5 @@
-import { NullableOptions, NumericStringSchema, OptionsForNumericString } from "../schemaClasses/NumericStringSchema.ts";
+import { NullableOptions } from "../libs/publicTypes.ts";
+import { NumericStringSchema, OptionsForNumericString } from "../schemaClasses/NumericStringSchema.ts";
 export { NUMERIC_STRING } from "../schemaClasses/NumericStringSchema.ts";
 export function numericString(options: OptionsForNumericString & NullableOptions): NumericStringSchema<null>;
 export function numericString(options: OptionsForNumericString): NumericStringSchema;
