+++ src/schemaClasses/EmailSchema.ts
@@ -8,8 +8,6 @@ import * as type from "../appliers/string/type";
 
 import {BaseSchema} from "./BaseSchema";
 
-export {NullableOptions} from "./BaseSchema";
-
 export type OptionsForEmail =
 	ifEmptyString.Options<string> |
 	ifNull.Options<string> |
