+++ src/schemaClasses/NumericStringSchema.ts
@@ -14,8 +14,6 @@ import * as type from "../appliers/string/type";
 
 import {BaseSchema} from "../schemaClasses/BaseSchema";
 
-export {NullableOptions} from "../schemaClasses/BaseSchema";
-
 export const NUMERIC_STRING = {
 	CHECKSUM_ALGORITHM: checksum.CHECKSUM_ALGORITHM,
 };
