+++ src/schemas/string.ts
@@ -1,4 +1,5 @@
-import {NullableOptions, OptionsForString, StringSchema} from "../schemaClasses/StringSchema";
+import {NullableOptions} from "../libs/publicTypes";
+import {OptionsForString, StringSchema} from "../schemaClasses/StringSchema";
 
 export {STRING} from "../schemaClasses/StringSchema";
